+++ src/parser.rs
@@ -34,6 +34,7 @@ pub(crate) mod go;
 mod helper;
 pub(crate) mod java;
 pub(crate) mod js;
+pub(crate) mod julia;
 pub(crate) mod kotlin;
 pub(crate) mod objective_c;
 pub(crate) mod python;